Move `toHumanReadableTime()` to DimenUtils
